requirements.txt
xmltodict==0.12.0
pyyaml==6.0